--- packages/react/src/close-button/package.json
@@ -1,4 +1,0 @@
-{
-    "sideEffects": true,
-    "name": "@cds/react/internal/close-button"
-  }
