+++ packages/react/src/internal-components/overlay/package.json
@@ -1,0 +1,4 @@
+{
+  "sideEffects": true,
+  "name": "@cds/react/internal-components/overlay"
+}
