+++ packages/react/src/internal-components/panel/entrypoint.tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../tsconfig.lib.json",
+  "compilerOptions": {
+    "composite": true
+  },
+  "references": [{ "path": "../../utils/entrypoint.tsconfig.json" }]
+}
