+++ packages/react/src/internal-components/panel/index.test.tsx
