+++ packages/react/src/internal-components/panel/index.tsx
@@ -1,7 +1,8 @@
 import { CdsInternalPanel as Panel } from '@cds/core/internal-components/panel';
+import '@cds/core/internal-components/panel/register';
 import { createComponent } from '@lit-labs/react';
 import * as React from 'react';
-import { logReactVersion } from '../utils/index';
+import { logReactVersion } from '../../utils/index';
 
 export const CdsInternalPanel = createComponent(React, 'cds-internal-panel', Panel, {}, 'CdsInternalPanel');
