+++ packages/react/src/internal-components/popup/entrypoint.tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../../tsconfig.lib.json",
+  "compilerOptions": {
+    "composite": true
+  },
+  "references": [
+    {
+      "path": "../../utils/entrypoint.tsconfig.json"
+    }
+  ]
+}
