+++ packages/react/src/internal-components/popup/package.json
@@ -1,0 +1,4 @@
+{
+  "sideEffects": true,
+  "name": "@cds/react/internal-components/popup"
+}
