+++ packages/react/src/internal-components/visual-checkbox/index.tsx
@@ -2,7 +2,7 @@ import { CdsInternalVisualCheckbox as VisualCheckbox } from '@cds/core/internal-
 import '@cds/core/internal-components/visual-checkbox/register';
 import { createComponent } from '@lit-labs/react';
 import * as React from 'react';
-import { logReactVersion } from '../utils/index';
+import { logReactVersion } from '../../utils/index';
 
 export const CdsInternalVisualCheckbox = createComponent(
   React,
