+++ packages/react/src/internal-components/visual-checkbox/package.json
@@ -1,0 +1,4 @@
+{
+  "sideEffects": true,
+  "name": "@cds/react/internal-components/visual-checkbox"
+}
