--- packages/react/src/panel/package.json
@@ -1,4 +1,0 @@
-{
-  "sideEffects": true,
-  "name": "@cds/react/panel"
-}
