--- packages/react/src/popup/package.json
@@ -1,4 +1,0 @@
-{
-    "sideEffects": true,
-    "name": "@cds/react/popup"
-  }
