--- packages/react/src/visual-checkbox/package.json
@@ -1,4 +1,0 @@
-{
-    "sideEffects": true,
-    "name": "@cds/react/internal/visual-checkbox"
-  }
