+++ packages/react/tsconfig.project.json
@@ -1,41 +1,113 @@
 {
   "extends": "./tsconfig.lib.json",
   "references": [
-    { "path": "./src/utils/entrypoint.tsconfig.json" },
-    { "path": "./src/accordion/entrypoint.tsconfig.json" },
-    { "path": "./src/alert/entrypoint.tsconfig.json" },
-    { "path": "./src/badge/entrypoint.tsconfig.json" },
-    { "path": "./src/breadcrumb/entrypoint.tsconfig.json" },
-    { "path": "./src/button/entrypoint.tsconfig.json" },
-    { "path": "./src/card/entrypoint.tsconfig.json" },
-    { "path": "./src/checkbox/entrypoint.tsconfig.json" },
-    { "path": "./src/close-button/entrypoint.tsconfig.json" },
-    { "path": "./src/datalist/entrypoint.tsconfig.json" },
-    { "path": "./src/date/entrypoint.tsconfig.json" },
-    { "path": "./src/divider/entrypoint.tsconfig.json" },
-    { "path": "./src/dropdown/entrypoint.tsconfig.json" },
-    { "path": "./src/file/entrypoint.tsconfig.json" },
-    { "path": "./src/forms/entrypoint.tsconfig.json" },
-    { "path": "./src/icon/entrypoint.tsconfig.json" },
-    { "path": "./src/input/entrypoint.tsconfig.json" },
-    { "path": "./src/modal/entrypoint.tsconfig.json" },
-    { "path": "./src/navigation/entrypoint.tsconfig.json" },
-    { "path": "./src/pagination/entrypoint.tsconfig.json" },
-    { "path": "./src/panel/entrypoint.tsconfig.json" },
-    { "path": "./src/password/entrypoint.tsconfig.json" },
-    { "path": "./src/popup/entrypoint.tsconfig.json" },
-    { "path": "./src/progress-circle/entrypoint.tsconfig.json" },
-    { "path": "./src/radio/entrypoint.tsconfig.json" },
-    { "path": "./src/range/entrypoint.tsconfig.json" },
-    { "path": "./src/search/entrypoint.tsconfig.json" },
-    { "path": "./src/select/entrypoint.tsconfig.json" },
-    { "path": "./src/selection-panels/checkbox/entrypoint.tsconfig.json" },
-    { "path": "./src/selection-panels/radio/entrypoint.tsconfig.json" },
-    { "path": "./src/tag/entrypoint.tsconfig.json" },
-    { "path": "./src/textarea/entrypoint.tsconfig.json" },
-    { "path": "./src/time/entrypoint.tsconfig.json" },
-    { "path": "./src/toggle/entrypoint.tsconfig.json" },
-    { "path": "./src/tree-view/entrypoint.tsconfig.json" },
-    { "path": "./src/visual-checkbox/entrypoint.tsconfig.json" }
+    {
+      "path": "./src/utils/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/accordion/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/alert/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/badge/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/breadcrumb/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/button/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/card/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/checkbox/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/datalist/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/date/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/divider/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/dropdown/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/file/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/forms/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/icon/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/input/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/modal/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/navigation/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/pagination/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/internal-components/close-button/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/internal-components/panel/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/internal-components/visual-checkbox/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/password/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/internal-components/popup/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/progress-circle/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/radio/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/range/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/search/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/select/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/selection-panels/checkbox/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/selection-panels/radio/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/tag/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/textarea/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/time/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/toggle/entrypoint.tsconfig.json"
+    },
+    {
+      "path": "./src/tree-view/entrypoint.tsconfig.json"
+    }
   ]
 }
